+++ src/parser/inline/autolink/account.ts
@@ -1,3 +1,4 @@
+import { undefined } from 'spica/global';
 import { AutolinkParser } from '../../inline';
 import { validate, focus, creator } from '../../../combinator';
 import { html } from 'typed-dom';
@@ -5,10 +6,16 @@ import { html } from 'typed-dom';
 // https://example.com/@user must be a user page or a redirect page going there.
 
 export const account: AutolinkParser.AccountParser = creator(validate('@', focus(
-  /^@(?:(?![a-z0-9.-]{0,200}?--)[a-z0-9](?:[a-z0-9-]{0,61}[a-z0-9])?(?:\.[a-z0-9]{1,63}){1,2}\/)?[A-Z-a-z][A-Za-z0-9]*(?:-[A-Za-z0-9]+)*/,
+  /^@(?:[A-Za-z0-9](?:[A-Za-z0-9-]{0,61}[A-Za-z0-9])?(?:\.[A-Za-z0-9](?:[A-Za-z0-9-]{0,61}[A-Za-z0-9])?)*\/)?[A-Z-a-z][A-Za-z0-9]*(?:-[A-Za-z0-9]+)*/,
   (source, { origin = '' }) => {
     const url = source.includes('/')
       ? `https://${source.slice(1).replace('/', '/@')}`
       : `${origin}/${source}`;
-    return [[html('a', { class: 'account', href: url, rel: 'noopener' }, source)], ''];
+    return verify(source) && [[html('a', { class: 'account', href: url, rel: 'noopener' }, source)], ''];
   })));
+
+function verify(source: string): true | undefined {
+  return source.length - ((source.indexOf('/') + 1) || 1) <= 64
+      && source.length - 1 <= 254
+      || undefined;
+}